Refactor: Consolidate imports in HeaderInput and HeaderInputRightSide components
