+++ src/widgets/Header/ui/HeaderInput.tsx
@@ -1,12 +1,15 @@
+'use client';
+
 import { ReactNode, type FC } from 'react';
-import { Input } from '@/shared/ui/Input/Input';
+import { Input, Container } from '@/shared/ui/index';
 import { SearchIcon } from '@/shared/ui/Icons/Icons';
-import { Container } from '@/shared/ui/Container/Container';
+import { useTranslation } from 'react-i18next';
 
 export const HeaderInput: FC<{ children: ReactNode }> = ({ children }) => {
+  const { t } = useTranslation();
   return (
     <Container className="flex w-full max-w-[500px] items-center gap-5 rounded-[30px] bg-white p-[10px]">
-      <Input size="md" placeholder="Search" leading={<SearchIcon />} />
+      <Input size="md" placeholder={t('HEADER.input_placeholder')} leading={<SearchIcon />} />
       {children}
     </Container>
   );
